django_requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
Pillow==10.0.1
stripe==6.7.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
django-storages==1.14.2
boto3==1.34.0
django-allauth==0.57.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-extensions==3.2.3
django-debug-toolbar==4.2.0
python-dotenv==1.0.0
pandas==2.1.3
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
django-subscription==0.1.0
django-plans==1.0.0